requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
pandas_ta>=0.3.14b
numpy>=1.24.0
scipy>=1.10.0
python-telegram-bot>=20.0
apscheduler>=3.10.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
